+++ js/screens/screen-stats/screen-stats.js
@@ -1,0 +1,7 @@
+import ScreenStats from '../screen-stats/screen-stats-view';
+
+export default (state) => {
+  const screenStats = new ScreenStats(state);
+
+  return screenStats;
+};
