+++ js/view.js
@@ -1,0 +1,24 @@
+import {getElementFromTemplate} from './engine/get-element-from-template';
+
+export default class AbstractView {
+
+  get template() {
+    throw new Error(`You have to define template for view`);
+  }
+
+  render() {
+    return getElementFromTemplate(this.template);
+  }
+
+  bind() {
+
+  }
+
+  get element() {
+    if (!this._element) {
+      this._element = this.render();
+      this.bind();
+    }
+    return this._element;
+  }
+}
